* Merged MIXER CROP, MIXER ANCHOR, MIXER ROTATION and MIXER PERSPECTIVE from 2.0 * Merged reporting of current MIXER values when no parameters are given. * Changed so that each vertex / texture coordinate pair in frame_geometry is encapsulated in a struct, to not require producers to keep track of indexes in an a array. * Changed so that the same code in image_kernel draws a single QUAD and a list of QUAD's * TODO make perspective correction and cropping work with non default geometry like the text producer. Currently no cropping or perspective correction works for custom geometry. * TODO make interaction coordinate translations work with new transforms. * Reverted text_producer change so that the baseline of the font is the actual position. * Fixed so that JSON encoded template data is also forwarded as is, instead of being interpreted as an .ftd file name. * Implemented anchor, crop, perspective and rotation support in scene_producer
Former-commit-id: 7cc9bc38e8e43e6f5b4511e5529a7faf46397d43

+++ core/producer/text/utils/texture_atlas.h
@@ -81,7 +81,7 @@ public:
 	texture_atlas(const size_t w, const size_t h, const size_t d);
 
 	rect get_region(int width, int height);
-	void set_region(const size_t x, const size_t y, const size_t width, const size_t height, const unsigned char *data, const size_t stride, const color<float>& col);
+	void set_region(const size_t x, const size_t y, const size_t width, const size_t height, const unsigned char *data, const size_t stride, const color<double>& col);
 	void clear();
 
 	size_t depth();
